Continue python source restructure
