--- ocf_devices/slined_onboarding/__init__.py
@@ -1,2 +1,0 @@
-from slined_onboarding.wpa_dpp_qr import get_dpp_uri, start_dpp_listen
-from slined_onboarding.soswitch import SoSwitch
