+++ ocf_devices/slined_onboarding/common/__init__.py
@@ -1,4 +1,4 @@
-from slined_onboarding.gui.so_pi import SoPiUi
-from slined_onboarding.gui.so_gpio import SoGpioContext
-from slined_onboarding.gui.so_img import SoImgLabel
-from slined_onboarding.gui.switch_main import gui_main
+from slined_onboarding.common.so_pi import SoPiUi
+from slined_onboarding.common.so_gpio import SoGpioContext
+from slined_onboarding.common.so_img import SoImgLabel
+# from slined_onboarding.common.wpa_dpp_qr import get_dpp_uri, start_dpp_listen
