--- ocf_devices/slined_onboarding/common/__main__.py
@@ -1,4 +1,0 @@
-from slined_onboarding.gui import gui_main
-
-if __name__ == '__main__':
-    gui_main()
