+++ ocf_devices/slined_onboarding/lightswitch/__main__.py
@@ -1,0 +1,4 @@
+from slined_onboarding.lightswitch import switch_main
+
+if __name__ == '__main__':
+    switch_main()
