+++ ocf_devices/slined_onboarding/lightswitch/switch_gui.py
@@ -1,0 +1,5 @@
+import logging
+from slined_onboarding.common import SoPiUi
+
+class SwitchUi(SoPiUi):
+    pass
